refactor: split GraphQL module "projects" resolvers into separate files
