+++ src/api/schema/bida/modules/projects/constants.ts
@@ -1,0 +1,1 @@
+export const DEPARTMENTS = ['FRONTEND', 'BACKEND', 'ANDROID', 'IOS']
